gitea_repos_master/drupe/drupe-mobile-api/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
pycryptodome==3.14.1
pydash==5.1.0
python-dotenv==0.19.2
requests==2.27.1
urllib3==1.26.8
